dt/transforms: Test transforms during partial upgrade
Part of the transforms RPC surface has been factored out into kafka::data with
a fallback to the old behavior subject to a feature flag.

This commit adds a ducktape test that upgrades from previous feature to HEAD
to demonstrate the fallback. Without it the test would fail as the produce
pass fails on rpc_not_found.

Signed-off-by: Oren Leiman <oren.leiman@redpanda.com>