update large_messages_test
This is an update to LMT to get it running again and test additional
scenarios.

Includes the following adjustments:

Adjust max.messages.bytes topic property based on the size of the
messages being tested: In the LMT we may test message sizes (hence
batches) which are larger than the default max batch size accepted by
Redpanda. Adjust the test so that it increases the topic property in
those cases only, and have it adjust to only a small amount higher than
the test message size (to account for batch and message headers).

This ensures that batches that we expect to work with default config
are tested at default config, and also that raising the setting works
as we expect.

Add MANY_PARTS mode which tests 1 topic and many partitions.

Other improvements.