bazel: build seastar in debug mode by default
Found out today that a benchmark I run was actually run weeks ago was
actually run in debug mode where I believed that we always run
benchmarks in release mode. This might be surprising expectation for you
but it wasn't for me as I'm used to the fact that we never build
benchmarks in debug with cmake at all.

Anyway...

When seastar is built in debug mode there is a helpful message

```
WARNING: debug mode. Not for benchmarking or production
```

and I was surprised I could ignore it. Turns out I didn't. It wasn't
printed because seastar wasn't being built in debug mode.

So I'm fixing that and making the bazel match more the cmake one.