r/consensus: do not reply with failure if follower batches match
When follower receives an append entries request with the vector of
records that are all present in its own log and their offsets and terms
match it should reply with success and correct `last_dirty_log_index`.
This way a leader instead of moving the follower `next_offset` backward
can start recovery process and deliver batches which the follower is
missing.

Signed-off-by: Michał Maślanka <michal@redpanda.com>