`kafka`: use `conditional_t` for floating point tristate
These call sites were previously hardcoded to `int64_t` for the
type used in a `boost::lexical_cast`.

With the advent of the first `tristate<double>`, this is no longer
a valid default.

Use a conditional type in order to account for the possibility of
a floating point tristate value conversion.