dt: demote allowlisted log line message to INFO
When we get an ERROR log line which is allowed (i.e., should not fail
the test), we output this info at WARN log level. WARN log lines during
dt runs go to the "session log", i.e., the log that appears on stdout
so this often spams the console and session log with tons of benign
messages about allowed log lines.

Demote this to INFO instead. It will still show in the test log.