ice/compat: Disallow dropping partition fields from schema
Though this restriction does not appear in the spec, dropping a data column
that also appears in the table's partition spec can cause validation errors
in clients, downstream. It may be possible to avoid this by performing live
partition spec reconciliation inline with the schema update, but for now we
simply reject such an update out of hand.

Also slightly refactors to consolidate error checks in validate_schema_transform.

Signed-off-by: Oren Leiman <oren.leiman@redpanda.com>