http/utils: changed url encoding to use upper case hex representation
According to the S3 documentation related with signature calculation the
canonical URI format must be encoded and use the upper case characters
in hexadecimal characters representation.

Document:
https://docs.aws.amazon.com/AmazonS3/latest/API/sig-v4-header-based-auth.html

This PR changes the canonical URI encode to use the upper case
hexadecimal representation.

The problem started to occur when a partition key was added as a part of
Iceberg parquet file path. As the path may contain some characters that
require encoding.

Signed-off-by: Michał Maślanka <michal@redpanda.com>