`rptest`: deflake `log_compaction_test`
These asserts were race-y due to the fact a segment roll could easily
add to the dirty/closed bytes of a log.

Furthermore, these asserts don't really prove anything, as we are waiting
for the number of dirty/closed bytes to eventually go to zero due to
retention enforcement anyways.