dl/coordinator: only load tables with pending state
Addresses a TODO in the iceberg_file_committer that highlighted that we
previously made some extraneous calls to the catalog.

The previous code structure relied on checking the STM state between
scheduling points to handle concurrent updates, and bailing out. This
meant that it was tricky to guarantee the presence or absence of DLQ
files during this method, and so we previously always attempted to load
both the main and DLQ tables.

Instead, this commit updates the concurrency strategy of this function
to make a copy of the STM state for the topic in question, and rely on
STM validation to ensure updates are rejected appropriately if there was
indeed a concurrent STM update that invalidated the new state change.

This change has the added benefit of avoiding potential iterator
invalidations that we've seen in this function in the past.