iceberg/rest_client: introduce client probe
Introduces a client probe to be used by the Iceberg REST client. For
now, this is just a wrapper around the HTTP client probe, but subsequent
commits will add additional metrics more specific to the catalog.

A label instance is included in this probe to differentiate the probe
used by the coordinator vs the one used by translators (without this,
the metrics subsystem don't like duplicate metrics registrations)