crash_tracker: filter out current crash report
Filter out the current process's crash reports from the result of
`recorder::get_recorded_crashes`.

This avoids misleading WARN messages when listing the current crash
reports while uploading crashes for telemetry. This is because the
current crash report file is empty at this point, which means that it is
currently seen as a malformed crash report file, and leads to a warning
log getting printed in `recorded_crashes_walker_fn`.

It also ensures that we don't even consider the current crash tracker
file for cleanup in `recorder::remove_old_crashfiles()`. The current
crash report is unlikely to be selected anyway because it should be the
newest file, but it is safer to exclude it outright from cleanup.