rpc_compiler.py: remove Codec template param
Every existing service always uses rpc::default_message_codec as its
protocol, and it has been that way for years as far as I can tell.

This change moves Codec template parameter into the class as a typedef
as, which means the service class is no longer a template, as a
precursor to some other changes to the rpc compiler.