dl/coordinator: don't remove snapshots if no topic state
Iceberg snapshot removal runs on every tick of the coordinator loop, and
unconditionally checks both the DLQ and main tables. At best, this is
just noise, loading tables that don't exist. At worst, this can pile up
and cause excessive traffic to the catalog.

This updates the snapshot remover to only run if there is pending
entries for the given table, since presumably we mostly care about
snapshot removal for tables that will get new snapshots when committing
new files.