k/quota_mgr: take now parameter explicitly
Fixes a class of bugs where delays in handling a request could cause
request parts to be throttled separately instead of as a unit.

For example, with a partition mutation quota limit of 10, a
create_partitions request adding 11 partitions to topic foo and 1
partition to topic baz should result in the baz part failing due to
quota limits. However, scheduling delays while updating the quota state
for foo could cause the baz request part to be processed later,
bypassing throttling. This inconsistent behavior is confusing
externally, so this commit ensures a single shared now timestamp is used
across a request, preventing such cases.

This fixes flakiness in the test_partition_throttle_mechanism test.