datalake/coordinator: group duplicate ensure_table calls
In certain scenarios with a large number of partitions I would see that
sometimes translation would get stuck, with RPCs to the coordinator all
timing out. I suspect that this is because of the large volume of
ensure_table_schema requests that could come in, particularly right
after a restart, when all translators are starting up together.

To mitigate this, this commit adds tracking of on-going requests to
ensure_table_schema and groups incoming requests in with on-going ones,
if they exist.