iceberg/rest_client: use /v1/config endpoint
Changes the iceberg_rest_catalog_prefix to
iceberg_rest_catalog_warehouse, with the expectation that this more
closely aligns with pyiceberg and other tools that allow configuring the
"warehouse".

This works by sending a request to the /v1/config endpoint with the
warehouse as a query parameter, and using the 'prefix' field returned by
the catalog, if any.

If no prefix is returned and a warehouse was specified, the warehouse
will used as a prefix instead as a fallback.

I've updated existing tests to use the new
iceberg_rest_catalog_warehouse instead of prefix. I also manually
verified this against a Unity Catalog, whose prefix was returned as
"catalogs/<warehouse>".