datalake: silence oom warning when finish is requested
If finish translation has been requested of a running translator then
currently preemption is provided by the stop translation interface which
unfortunatley in its current form implies an out-of-memory condition
which logs a warning. This commit silences the warning when a translator
is preempted and finish translation has also been requested.

Once a generalized, more expressive interface can be used to
finish/stop/preempt a translator some of these flag checks/hacks can be
removed.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>