`http`: add `utils.cc/.h` and `flexible_function_handler`
`ss::httpd::function_handler` calls `rep->done(_type)` on the reply before
returning it. Because certain code paths in imposter services may want to
return a different `Content-Type` in their response, this impl's `handle()`
function allows for modifying the `type` set in the function handler's body
before the final call to `done(type)` is performed.

This class also handles an oversight in `mime_types.cc`, in which `xml` is
not handled as a possible `mime` type.