benchmarks: run benchmarks in shared memory
Introduce bench-wrapper.sh, which is a wrapper script for a
microbenchmark. The main thing it does now is runs the benchmark in
shared memory. This is important because the benchmark may
do IO and doing it to shm is low-variance and fast. The old ctest way
did it this way and this resulted in spurious CDP detection when we
moved to bazel.

To run the benchmark int the wrapper, append `.run` to the benchmark
target, like:

```
bazel run //src/v/kafka/server/tests:kafka_produce_partition_rpbench.run
```

The benchmark script also optionally redirects stderr to a file. This
defaults to off but can be enabled on a per-bench basis when calling
the cc_bench macro, or by the user in the environment by setting
MB_REDIRECT_STDERR=1.

Fixes CORE-9760.