test: fix datalake space usage test
The primary issue here after fixing the memory accounting in serde
parquet reader is that the 10% shard memory limit is reached very
quickly in ducktape where we are memory constrained in CI. This means
that translators frequently finish and remove their data from disk.

The fix is (1) give ducktape more memory (more memory and less cores),
and (2) test with larger number of partitions which has the effect of
making it less likely that any one partition translator will reach the
finish limit. Still, it's a bit fuzzy and may need additional tuning.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>