r/stm_manager: fixed bug in stm manager leading to unnecessary reads
The state machine manager should select the max offset from all the
available so the lagging state machines can be caught up in the
background apply. Previously the stm manager selected the smallest
offset leading to catching up the state machines in foreground and
trying to call apply on the state machines which are already up to date.

Signed-off-by: Michał Maślanka <michal@redpanda.com>