r/state_machine: introduced initial recovery policy
Introduced the `stm_initial_recovery_policy`. The policy is there for
the state machine implementer to decide how to handle the situation when
a state machine is crated for the first time. The state machine can
choose from two options:

`read_everything` - the read everything policy simply creates an stm and
apply each batch of the whole locally available log. The application is
being done in the `background` apply fiber of state machine manager
until the stm is fully caught up. (This was the default behavior)

`skip_to_end` - this policy allows the state machine not to apply the
whole log when it is created for the first time. The skip to the end
policy starts state machine initial recovery from the very specific
offset. The offset is established as the next to highest last applied
offset returned from all the state machines created on the same
partition previously.

Why we can not skip to the log end offset with `skip_to_end` ?

We can not skip to the end offset of the log as it may have not been
committed and therefore can be truncated. In this situation the stm
could loose some applies as its `_next` wouldn't change with the
truncation and only batches with offset greater than `_next` would be
applied to the state machine. Using last applied offset from other state
machines eliminates this problem as last applied offset must have been
committed before to be applied.

Signed-off-by: Michał Maślanka <michal@redpanda.com>