`storage`: consider `uint32_t` offset space in adjacent compaction
Ensure that the adjacent segments do not span an offset space greater
than the maximum value that can be represented by a uint32_t. This must
be enforced due to use of `roaring::bitmap` in the `compacted_offset_list`,
which is used to deduplicate records during self compaction and sliding
window compaction. Overflows in this area could lead to incorrect compaction.