bazel: Fix OpenSSL build to get correct openssldir
When OpenSSL looks for system certificates, it uses a path relative to
the macro "OPENSSLDIR".  This is compiled into the OpenSSL library and
is set when the build is configured.

The OpenSSL Configure script takes in `--openssldir` which sets the
OPENSSLDIR macro to the path specified, relative to the value set to
`--prefix`, so the OPENSSLDIR macro becomes `<prefix>/<openssldir>`.

In most Linux systems, the system certificates are installed in
`/etc/ssl/certs`, so OpenSSL needs to have its OPENSSLDIR macro set to
`/etc/ssl`.  To do this in bazel we have to do the following:

* Set `--prefix` to `/`
* Set `--openssldir` to `/etc/ssl`
* Run `make` with `DESTDIR` set to the directory where the build products
  should be installed into.

This arised when errors were observed in the OIDC ducktape tests when
using TLS.  In ducktape, we install root certificates into
/etc/ssl/certs, however because the OpenSSL build was not looking for
the system certificates there, the test would fail because OpenSSL did
not trust our root CA certificate.  This change addresses this issue.

Signed-off-by: Michael Boquard <michael@redpanda.com>