dt/redpanda: added usage stats to redpanda service
Added usage statistics to Redpanda service. The statistics are going to
be reported in the test result and used to detect anomalies.

Some of the changes introduced in Redpanda core may have a large impact
on Redpanda resource consumption for example an amount of data read or
written. The tests usually do not validate the resource usage but rather
the feature correctness. Collecting information about the cluster usage
for every test that runs will allow us to develop an independent anomaly
detection tool that will warn us if the resource use will be
significantly different than before.

Signed-off-by: Michał Maślanka <michal@redpanda.com>