dl/schema: avoid using blanket parquet_conversion_error
Currently parquet_conversion_error is the blanket error type visible in
the top level translator (and hence the logs) for any error in type
resolution (eg: schema registry is down) or in dlq paths. The latter is
rare because dlq schema is tightly controlled while the former is pretty
common.

The commit tries to decouple these and reserve parquet_conversioni_error to
paths where low level parquet conversion is indeed the problem.