cluster: add `errc::topic_id_already_exists`
This will be returned from command apply methods on topic creation and
topic id assignment if the given topic id already exists.

This error should not really happen (assuming the uuids used for topic
ids are always unique), but it is safer to handle this as an internal
error instead of asserting that they never happen.