archival: Rejigger segment collector concurrency control
For non-reupload segment collection, it is not uncommon to arrive in a state
where the tail of a collected segment list is an unsealed segment. This
generally means that time based upload was triggered, and we are effectively
forcing this segment up to s3.

This is desirable behavior in general, but it has an undesirable interaction
with concurrency control code in make_upload_candidate. Since the tail segment
is open, there is a greater chance that it's generation ID will be incremented
(e.g. by some append) between discovery and read-lock acquisition, leading to
the entire upload being discarded. This happens more frequently than, for
example, compaction kicking in on already collected segments and can lead to
lots of upload candidates being discarded.

This commit adds a special case to existing concurrency control code for
exactly this situation. That is, we skip the generation ID check if and only if

- the last segment in the list was NOT closed AND
- the last segment is the ONLY segment with gen ID difference

In this case, it should be safe to assume that compaction hasn't taken place
and that, once we have acquired read locks, the size of the collected segments
on disk will not have changed. Therefore we proceed with the upload.

In any other case where there is a generation ID mismatch, we short circuit
the upload as usual.

Signed-off-by: Oren Leiman <oren.leiman@redpanda.com>