`storage`: check for hash key map of capacity 0 in `do_compact()`
Allowing a `compaction_config` with a `hash_key_map` of capacity 0 through
to sliding window compaction results in a floating point exception core dump.

This means users who:

1. Start with a cluster config of `log_compaction_use_sliding_window=false`
2. Execute `rpk cluster config set log_compaction_use_sliding_window=true`

Can easily run into a crash.

Prevent the crash by adding a new check for the capacity of the `hash_key_map`
and fall back to adjacent segment compaction if it is 0.