`config`: reduce `max_compacted_log_segment_size` default
This default value of `5_GiB` is quite massive and was chosen long ago.
With the advent of changes resulting in more aggressive adjacent merging
during compaction, this default needs revisiting to ensure the new behavior
won't cause stability issues in the extreme case where a long running `log`
suddenly has its average `segment` size trend towards
`max_compacted_log_segment_size` due to compaction.

Our cloud tier config generators have dropped the default down to twice
the value of `log_segment_size` [1].

A rational explanation for why twice the value of `log_segment_size` makes
sense for this default is that for most `compact` enable topics with
a bounded keyset, historical `segment`s existing in the `log` should be
reduced to very small sizes as sliding window compaction works over the
entirety of the key-offset space. These `segment`s will be ostensibly be
reduced in number and increased in size, trending towards `max_compacted_log_size`.
Therefore, capping at `512_MiB` with the new default seems to be a safer
bet than the much larger than average (with regards to the statistics we
see in current `redpanda` clusters) `segment` size of `5_GiB`.

In the pathological case that sliding window compaction cannot reduce two
adjacent segments, twice the value of `log_segment_size` still allows for
the segments to be adjacently merged, so even in the worst case scenario,
the growth rate for the `segment` count in a `log` can be kept somewhat
bounded.

No `legacy_default` is set here, because its highly unlikely we ever came close
to reaching this `5_GiB` limit in previous clusters via adjacent segment
compaction, and because it is preferable to not retain this default for older
clusters that will upgrade to a new version with the hyper aggresive adjacent
merging.

[1]: https://github.com/redpanda-data/cloudv2/blob/940cfd4f3560b21df6d763faabf529c88a2546ef/install-pack/generate-tiers.py#L562)