selftest: fallocate files in selftest
fallocate in addition to ftruncate the test files. This is to mirror the
segment_appender.

Note in practice this doesn't make much difference as the test will
relatively quickly enter into the regime where the file is fully
allocated anyway and we are just overwriting anyway (all tests only use
a 1GB total file size currently).