`ssx`: modify `async_transform()` signature & return type
This was hardcoded to return a `std::vector<result_type>`. We may not
always want this container type as our return value.

Modify the signature to take a `template<typename ContainerType>`, allowing
the caller to specify the container which should be returned.

Also moves the concept `CanReserve` to `base/type_traits.h`, which is used
to determine if space can be reserved up front in `async_transform()`.