k/client: reduce retries for client::list_offsets
Do not retry non-retriable errors, unless an external mitigator is
configured.

When no external mitigator is configured, this reduces the
partition-level errors we retry on, from all error codes to only
retriable ones.
When an external mitigator is configured, the retry behaviour is as
before. We retry all partition-level errors.

This is to reduce retries in the default case, when no external
mitigator is configured. When an external mitigator is configured, we
keep the existing behaviour of retrying on all errors since the external
mitigator might be able to mitigate it.

We also introduce `throw_on_error` to ensure that the throwing behaviour
remains as before, that is, `list_offsets` throws on any kafka error in
the response.

When no mitigator is configured, the client behaviour attempts to mimic
franz-go's retry behaviour on list_offset responses:
https://github.com/twmb/franz-go/blob/12af19728d2ea6b93fbdb26cbaeb1b2d7208e668/pkg/kgo/client.go#L2921-L2925