rptest/cluster_restore_test: measure progress in records
Updates cases of cluster_restore_test with slow tiered storage (where we
expect to have duplicate offsets) that previously waited for the max
offset to increase as a signal for ingestion progress. Instead, we now
wait for the count of the table to go up, which is a much faster
condition to achieve (e.g. otherwise the test may have to effectively
have to wait for the table to double in size).