`storage`: change return type of `do_self_compact_segment()` to `compaction_result`
No functional changes here, `self_compact_segment()` returns a `compaction_result`
using the same data that could be used to build the object at a layer down
in the call stack.

This will also have the added benefit of allowing the `cmp_idx_size_after`
to be passed back to callers of `self_compact_segment()`.