c/a/upload: Adds timestamps & compaction status to reconciliation_result
Note that this code isn't currently used anywhere but provides the building
blocks for constructing remote segment uploads through a log reader interface.

In the current implementation we have individual segments in hand, so we can
look directly at things like compaction status and batch timestamps.

When going through a log_reader, direct log & segment manipulation is pushed
to a lower layer, but we still need to transmit this information up to archival
to populate segment metadata.

Signed-off-by: Oren Leiman <oren.leiman@redpanda.com>