`rptest`: add `RedpandaService.get_version_if_not_head()`
Returns the `redpanda` binary version as a string if it differs from HEAD.

The intention behind this function is to help add more context to `BadLogLines`
summaries during a failed test in future commits.

The motivation for this is that it is not always evident from a failed test's
summary whether the failure is present in the current `HEAD` of `redpanda`,
or if a failure has occured on a broker running a previous version. By adding
more context, it will be obvious on first glance (especially within CI summary
pages) whether a CI failure is a present concern or if a previous version
of `redpanda` is perhaps lacking known fixes, backports, etc.