ct/l1: use fixed header as the prefix delimiter
C++23 is so fun. Instead of using serde and a prefix, we instead write
the header as a fixed width and use the record batch size there to read
the batch data.

We use some fun templates to ensure reading and writing process the
right fields in the right order. Note that we can't share the same code
with local storage because local storage does not store the term
information in the batch header (That's in the filename instead, and
patched in after).