rptest: adjust cluster metadata upload frequency
The test was timing out while waiting for the cluster to have metadata
in S3 because each check was hitting a warning:

[DEBUG - 2025-06-27 06:41:36,069 - si_utils - _load_cluster_metadata_manifest - lineno:1199]: Exception loading cluster_metadata/ac55dd59-f0bc-4195-8b3f-2ca54470d8ae/manifests/36/cluster_manifest.json: An error occurred (NoSuchKey) when calling the GetObject operation: The specified key does not exist.

This is because the upload loop cleans up old metadata on each round of
uploading. To mitigate, this updates the upload frequency to happen less
often to give the test more time to see metadata.

Adjusts some timeouts accordingly to give a bit more wiggle room for the
slower uploads.