tests: fixed pattern used to detect log about stuck reconciliation
Partition balancer test may suspend the node to make it unavailable.
This way of unavailability is important to be covered in tests as it is
often triggering various race conditions. In this case it is expected to
see logs about stuck reconciliation as the node was suspended for a time
long enough to trigger the log.

Signed-off-by: Michał Maślanka <michal@redpanda.com>