`storage`: split compaction logic into `is_compactible()` & `is_filterable()`
The major motivation for this change is the fact that transactional batches
will be considered "filterable" in the future, despite not being "compactible".

After the changes in this PR, the property "compactible" refers to the fact
that a record may be removed due to a record with the same key existing later
in the log (de-duplication). "Compactible" records should be indexed in the
`.compaction_index` file for a `segment`, and they should be considered in
`should_keep::is_latest_key()`.

On the other hand, the property "filterable" means that a record may be removed
due to either de-deduplication OR by other means during the compaction process.
Most notably, the removal of transaction batches past the horizon set by
`delete.retention.ms`. Filterable batches should be processed by `should_keep()`,
but some filterable batches should not be considered in
`should_keep::is_latest_key()` (aforementioned transaction batches which
should only be removed by `delete.retention.ms`, and _not_ de-duplication).

The property `num_compactible_records_appended` in the `index_state` now refers
to filterable records, not just compactible records. However, the naming
throughout the `storage` layer for this function and its related call sites
are left the same to prevent this slight implementation detail/nuance from
leaking out too far outside of compaction.