ct: Propagate L0 metadata in fetch request
The cloud_topic_partition now propagates the metadata from the L0
metadat batch (dl_placeholder) to the raft_data batch returned to the
client in fetch request.

The placeholder batch contains the correct offset and term and some
other information required by the protocol. The data uploaded to S3 in
L0 objects contains record batch headers but these headers are not
initialized properly because the offsets are not allocated at the moment
when the L0 object is uploaded. The offsets are allocated when the
placeholder batches are replicated. The cloud_topic_partition now
combines this information with the user data in L0 objects to respond to
the fetch request.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>