rp/admin/server: Fix an issue with the audit log escape hatch
Audit log's escape hatch was not working and the full url would not
match against the endpoint path.

The get_url method we were using includes protocol name, header and url.
The url also includes the query parameters. To fix this, the _url method
is being used directly and the query parameters are stripped.

Also added the get_cluster_config endpoint in the allowed_requests,
which should have been there.