audit: Integrate use of the reject policy flag
This change integrates the use of the audit_log_reject_policy
configuration by allowing audit log messages to be dropped if they are
unable to be enqueued.  This could be because:

* auth was not properly configured on Redpanda to permit the audit log
  client to use ephemeral credentials, or
* the audit log buffers have become full

Additionally, a timeout is added for the when the application lifecycle
event is published.  This event records when Redpanda starts/stops or
when auditing is enabled/disabled.  This timeout is set to 5s if the 'drop'
policy is selected.  This is to ensure that, if the message is unable to be
produced because the audit log buffers are full, this does not hold up
Redpanda startup/shutdown.

Signed-off-by: Michael Boquard <michael@redpanda.com>