ct/l1: metastore replace_objects interface
Adds an interface to the metastore that allows callers to replace
extents with new extents that point to new objects.

Note, this commit only includes the interface for a simple replacement
of extents with no constraints other than offset boundary checks. This
may be used during leveling, but is not sufficient for compaction
without additional accounting, to come in a later change.