archival: Start compacted reupload conditionally
Currently, the compacted reupload is started during every upload cycle
if the topic is compacted.
This means that the segment_collector is created and used to find the
reupload candidate. But the entire log is compacted and uploaded we will
get an error "Start offset XXX is ahead of manifest last offset". This
error is logged on "WARN" but it doesn't affect reliability in any way.

To avoid this problem this commit adds additional condition for the
compacted reupload. In order for the reupload to be scheduled the last
uploaded compacted offset should lag behind the last uploaded offset.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>