tx/tm_stm: consolidate cleanup of _pid_tx_id map
The invariant is that _transactions and _pid_tx_id map is 1:1. This
change ensures to remove any existing mapping from _pid_tx_map before
isntalling new entries.

Without this change, it can happen that existing tx_it.pid's removal
maybe skipped in some cases.

Do note that last_pid purpose is something different (read kip-320) and
is not set in all cases (it may be a separate but in itself that needs
more eyes).