`cloud_storage_clients`: default `minio` backend url style to `path`
`virtual_host` style is only supported by MinIO iff the property
`MINIO_DOMAIN` is set [1]. If it is not, API calls will fail- with the exception
of `ListObjects`, which is the request type used to self configure the `s3_client`
in the case that `cloud_storage_url_style` is not set.

This leads to a bad state in which `MinIO` will continue to fail until the user
overrides `cloud_storage_url_style` manually to `path`.

To prevent this bad self-configuration, handle the `minio` backend case
explicitly by hard-coding its self configuration result to `path`.

If a user feels strongly about the `cloud_storage_url_style` being used,
they can always override it at the cluster level.

[1]: https://min.io/docs/minio/linux/reference/minio-server/settings/core.html#envvar.MINIO_DOMAIN